% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplt.R
\name{biplt}
\alias{biplt}
\title{make a biplot that allows you to see the relationship between two components of a PCA, the weight of the respective variables and to color by category}
\usage{
biplt(
  pca,
  choices = c(1, 2),
  col = "blue",
  threshold = 0.25,
  xlab = "cps",
  ylab = "cps",
  cex = 2,
  pch = 16,
  labels = rownames(pca$x),
  main = "",
  legend = FALSE,
  legend.labels = NULL,
  legend.pos = "topright",
  legend.pch = NULL,
  legend.title = "",
  legend.col = NULL
)
}
\arguments{
\item{pca}{a list with class "prcomp" from prcomp()}

\item{choices}{a vector indicating the indices of two desired components that we wish to relate}

\item{col}{a vector of color or a factor}

\item{threshold}{the elimination threshold to avoid displaying the variables that have the least influence on the components}

\item{xlab}{label of x}

\item{ylab}{label if y}

\item{cex}{length of points}

\item{pch}{pitch type}

\item{labels}{labels for the points}

\item{main}{title of the plot}

\item{legend}{logical, if TRUE display a legend}

\item{legend.labels}{vector of labels for the legend (if NULL, automatically deduced from col)}

\item{legend.pos}{position of the legend (default "topright")}

\item{legend.pch}{pch for the legend (if NULL, uses the same as pch)}

\item{legend.title}{title of the legend}

\item{legend.col}{vector of colors to use in legend (if NULL, automatically deduced)}
}
\value{
a graphic biplot more esthetic of biplot() function
}
\description{
make a biplot that allows you to see the relationship between two components of a PCA, the weight of the respective variables and to color by category
}
\examples{
# Example 1
data(iris)
a <- prcomp(iris[,1:4])
biplt(a, col=iris$Species, threshold=0.8, choices=c(1,3), legend=TRUE)

# Example 2 with custom colors
cols <- rainbow(3)[as.numeric(iris$Species)]
biplt(a, col=cols, threshold=0.8, legend=TRUE, legend.labels=levels(iris$Species))
}
