% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootreg.R
\name{bootreg}
\alias{bootreg}
\title{Validating a bootstrap linear regression model}
\usage{
bootreg(
  reg,
  data = c(),
  plot = TRUE,
  verbose = TRUE,
  conf.level = 0.95,
  alpha = 0.05,
  iter = 1000
)
}
\arguments{
\item{reg}{linear model.}

\item{data}{a data.frame of values if formula is complex with I() function (Optional).}

\item{plot}{enable or disable the display of graphical analysis.}

\item{verbose}{enable or disable the display of the commented analysis.}

\item{conf.level}{confidence level for validation of the model.}

\item{alpha}{minimal value accepted for validation of the model and his coefficients.}

\item{iter}{number of iterations.}
}
\value{
A graph of analysis.
}
\description{
Validating a bootstrap linear regression model
}
\details{
bootreg allows to validate a model by bootstrap.

It will draw several times in the values of the model in order to test its robustness.

As an output, we can have a validation or not of the model (analysis=F), or we can have a table expressing the variability of the coefficients, their p-values and their maximum fluctuation values.
}
\examples{
data(mtcars);
corrigraph(mtcars);
reg<- lm(cyl~disp+hp,data=mtcars);
bootreg(reg, verbose=TRUE, plot=TRUE)
}
