% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks.distance_lmer.R
\name{cooks.distance_lmer}
\alias{cooks.distance_lmer}
\title{Function to compute Cook's distance for mixed models}
\usage{
cooks.distance_lmer(model, obs = FALSE)
}
\arguments{
\item{model}{A mixed-effects model of class lmerMod (fitted with lme4).}

\item{obs}{Logical, if TRUE, computes Cook's distance for individual observations as well.}
}
\value{
A named vector with Cook's distance values for each group of random effects and optionally for each observation.
}
\description{
Function to compute Cook's distance for mixed models
}
\examples{
# Example 1
library(lme4)
data("sleepstudy")
model <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
cooks_distances <- cooks.distance_lmer(model)

# Example 2
data(mtcars)
model <- lmer(cyl ~ disp + (1 | hp) + (1 | gear), data = mtcars)
cooks_distances <- cooks.distance_lmer(model)
}
