% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.test.R
\name{.n_levels_g}
\alias{.n_levels_g}
\title{Calculate the number of levels in grouping variable(s)}
\usage{
.n_levels_g(g)
}
\arguments{
\item{g}{A factor, vector, or data.frame of factors}
}
\value{
Integer representing the number of levels (or combinations for data.frame)
}
\description{
Calculate the number of levels in grouping variable(s)
}
\note{
For data.frames, only categorical/factor variables are counted.
Numeric variables (covariates) are excluded to avoid false positives
in "too many groups" checks when ANCOVA models are used.
}
\keyword{internal}
