% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_one_factor_analysis.R
\name{.one_factor_analysis}
\alias{.one_factor_analysis}
\title{Perform One-Factor Analysis}
\usage{
.one_factor_analysis(
  x = NULL,
  g = NULL,
  formula = NULL,
  data = NULL,
  paired = FALSE,
  id = NULL,
  alpha = 0.05,
  return = TRUE,
  k = NULL,
  code = FALSE,
  debug = FALSE,
  verbose = FALSE,
  boot = TRUE,
  silent = TRUE
)
}
\arguments{
\item{x}{Numeric vector. Dependent variable to be analyzed.}

\item{g}{Factor. Grouping variable for the analysis.}

\item{formula}{Optional. A formula specifying the model (e.g., x ~ g).}

\item{data}{Optional. A data frame containing variables specified in the formula.}

\item{paired}{Logical. Indicates whether data are paired (default is FALSE).}

\item{id}{Optional. Identifier for paired data.}

\item{alpha}{Numeric. Significance level for hypothesis testing (default is 0.05).}

\item{return}{Logical. If TRUE, returns results; if FALSE, only displays output (default is TRUE).}

\item{k}{Optional. Counter for verbose messages.}

\item{code}{Logical. If TRUE, displays R code snippets (default is FALSE).}

\item{debug}{Logical. If TRUE, detailed debugging messages are displayed (default is FALSE).}

\item{verbose}{Logical. If TRUE, provides detailed output during the analysis (default is FALSE).}

\item{boot}{Logical. If TRUE, enables bootstrap-based procedures (default is TRUE).}

\item{silent}{Logical. If TRUE, suppresses warnings from underlying functions (default is TRUE).}
}
\value{
Returns results of the selected statistical tests, including:
\itemize{
\item p-values for hypothesis testing.
\item Results of normality checks and variance assessments.
}
}
\description{
This function performs statistical analysis for one grouping factor.
It handles both normal and non-normal data, adjusting the statistical tests accordingly.
}
\details{
The function performs the following:
\itemize{
\item Adjusts the significance level using Sidak's correction.
\item Tests normality of data using Shapiro-Wilk or Jarque-Bera tests, based on sample size.
\item Applies appropriate statistical tests: t-test, Mann-Whitney, ANOVA, Kruskal-Wallis.
\item Handles cases with equal or unequal variances and non-normal data.
}
}
\examples{
\dontrun{
# Example with normal data
set.seed(123)
x <- c(rnorm(50, mean = 5), rnorm(50, mean = 6))
g <- factor(rep(1:2, each = 50))
.one_factor_analysis(x, g, verbose = TRUE)

# Example with non-normal data
x <- c(rpois(50, lambda = 5), rpois(50, lambda = 7))
g <- factor(rep(1:2, each = 50))
.one_factor_analysis(x, g, verbose = TRUE)
}

}
\seealso{
\itemize{
\item \code{\link[fda.usc:fanova.hetero]{fda.usc::fanova.hetero()}}
\item \code{\link[agricolae:kurtosis]{agricolae::kurtosis()}}
\item \code{\link[lawstat:levene.test]{lawstat::levene.test()}}
\item \code{\link[WRS2:med1way]{WRS2::med1way()}}
\item \code{\link[onewaytests:bf.test]{onewaytests::bf.test()}}
}
}
\keyword{internal}
