% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{dsc}
\alias{dsc}
\title{Obtain the best configuration to meet the objectives determined by one or more linear models.}
\usage{
dsc(
  data,
  reg,
  Y = c(),
  ymin = c(),
  ymax = c(),
  pop = iter/20,
  iter = 4000,
  wash = pop/2,
  plot = T,
  verbose = F,
  save = F,
  file = "file.html"
)
}
\arguments{
\item{data}{A data.frame with X(s) and Y(s).}

\item{reg}{A linear model or a list of linear models.}

\item{Y}{Values that we want to achieve for the different Y predicted using the model (s).}

\item{ymin}{List of minimum values tolerated for the different Y.}

\item{ymax}{List of maximum values tolerated for the different Y.}

\item{pop}{Population of parameters which will cross randomly to generate better parameters.}

\item{iter}{Number of iterations in the scalable approach (should ideally be much greater than the popupulation (pop) of settings.}

\item{wash}{The maximum number of desired settings.}

\item{plot}{If TRUE, displays interactive parallel coordinates (plot_ly) to identify the best possible settings.}

\item{verbose}{If TRUE, gives information about the analysis.}

\item{save}{For saving the graph (html format)}

\item{file}{Name of the html page "xxx.html"}
}
\value{
A dataframe containing all the selected settings sorted from best (top) to worst (bottom).
}
\description{
Obtain the best configuration to meet the objectives determined by one or more linear models.
}
\examples{
\donttest{
data(mtcars)
reg1 <- lm(mpg ~ wt + qsec, data = mtcars)
reg2 <- lm(cyl ~ disp, data = mtcars)
reg <- list(reg1, reg2)
output <- dsc(mtcars, reg, Y = c(23.4, 5.4), pop = 50, iter = 200, plot = FALSE)
# Aggregation of several trials
for (i in 1:3) {
  output <- rbind(output, dsc(mtcars, reg, Y = c(23.4, 5.4), plot = FALSE))
}
parco(output, "Distance")
# With filtration of min and max y.
output <- dsc(
  mtcars, reg, Y = c(15, 5), ymin = c(14, 4), ymax = c(15, 6),
  pop = 50, iter = 200, plot = FALSE
)
}
}
