% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc2.R
\name{dsc2}
\alias{dsc2}
\title{Check the robustness of a suggested dsc() prediction of X for objectives in Y by bootstrapping}
\usage{
dsc2(data, reg, dsc, iter = 500, plot = TRUE, return = FALSE)
}
\arguments{
\item{data}{A data.frame with X(s) and Y(s).}

\item{reg}{A linear model or a list of linear models.}

\item{dsc}{A dataframe line containing the X's to be predicted (ideally a dsc() output).}

\item{iter}{Number of iterations in the scalable approach (should ideally be much greater than the popupulation (pop) of settings.}

\item{plot}{If TRUE, displays interactive parallel coordinates (plot_ly) to identify the best possible settings.}

\item{retun}{If TRUE, return the data.frame of values predicted by bootstrapping.}
}
\value{
A dataframe containing all the selected settings sorted from best (top) to worst (bottom).
}
\description{
Check the robustness of a suggested dsc() prediction of X for objectives in Y by bootstrapping
}
\examples{
data(mtcars)
colnames(mtcars)
myreg1 <- evolreg(mtcars,"mpg")
myreg2 <- evolreg(mtcars,"cyl")
reg <- list()
reg[[1]] <- myreg1
reg[[2]] <- myreg2
output <- dsc(mtcars,reg,Y=c(23.4,5.4),pop=400,iter=200)
# Bootstrap on the best line with dsc2()
output2 <- dsc(mtcars,reg,dsc=output[1,],plot=TRUE)
}
