% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvar.R
\name{dvar}
\alias{dvar}
\title{Diversifies the variables of a dataframe by testing interactions, polynomials, logs... so that evolreg can draw a larger number of model combinations.}
\usage{
dvar(
  data,
  Y,
  X = c(),
  alpha = 0.05,
  family = "lm",
  wash = TRUE,
  NAfreq = 1,
  interaction = TRUE,
  multix = TRUE,
  multidiv = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a dataframe.}

\item{Y}{the y to predict.}

\item{X}{variables whose presence we want to force in the model.}

\item{alpha}{0 to 1. If there are too many variables and the argument wash=TRUE, use this p-value threshold to eliminate the variables whose effect is too insignificant (Risk of eliminating the variables that will have an effect once transformed or in interaction).}

\item{family}{"lm", "logical" or "lmer". Type of regression}

\item{wash}{TRUE or FALSE.To select the best variables when there are too many.}

\item{NAfreq}{from 0 to 1. NA part allowed in the variables. 1 by default (100\% of NA tolerate).}

\item{interaction}{FALSE or TRUE. To allow interactions between variables.}

\item{multix}{FALSE or TRUE. To allow variable variants (log, exp, polynomial, ^2).}

\item{multidiv}{FALSE or TRUE. To allow the synthesis of variables combining the ratio of one variable divided by another.}

\item{verbose}{With an operating report.}
}
\value{
The dataframe of the data and a list of interaction formulas and transformations with associated p-values in their ability to predict Y.
}
\description{
Diversifies the variables of a dataframe by testing interactions, polynomials, logs... so that evolreg can draw a larger number of model combinations.
}
\examples{
data(iris)
dvar(iris,"Sepal.Length")
}
