% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolreg.r
\name{evolreg}
\alias{evolreg}
\title{Identify the best linear, logistic or mixed regression model using an evolutionary approach.}
\usage{
evolreg(
  data,
  Y,
  X = c(),
  pval = 0.05,
  nvar = 0,
  iter = 4000,
  multix = TRUE,
  interaction = TRUE,
  multidiv = TRUE,
  nbind = c(),
  wash = TRUE,
  NAfreq = 1,
  family = "lm",
  plot = FALSE,
  verbose = TRUE,
  fast = TRUE
)
}
\arguments{
\item{data}{a dataframe.}

\item{Y}{the y to predict.}

\item{X}{vector of variables whose presence we want to force in the model.}

\item{pval}{0 to 1. If there are too many variables and the argument wash=TRUE, use this p-value threshold to eliminate the variables whose effect is too insignificant (Risk of eliminating the variables that will have an effect once transformed or in interaction).}

\item{nvar}{Maximum number of variables in the model. A default value is proposed according to the number of individuals.}

\item{iter}{Number of iterations.}

\item{multix}{FALSE or TRUE. To allow variable variants (log, exp, polynomial, ^2).}

\item{interaction}{FALSE or TRUE. To allow interactions between variables.}

\item{multidiv}{FALSE or TRUE. To allow the synthesis of variables combining the ratio of one variable divided by another.}

\item{nbind}{Number of simulated individuals in the population of models to be crossed in an evolutionary approach. A default value is proposed according to the number of variables.}

\item{wash}{TRUE or FALSE.To select the best variables when there are too many.}

\item{NAfreq}{from 0 to 1. NA part allowed in the variables. 1 by default (100\% of NA tolerate).}

\item{family}{"lm", "logical" or "lmer". Type of regression}

\item{plot}{To visualize the evolution of the R2 of the models obtained after each crossing.}

\item{verbose}{To display a summary of the intermediate models.}

\item{fast}{Paramètre qui stoppe les itérations lorsque le R carré des modèles n'évolue plus de façon significative.}
}
\value{
A strongest possible regression model chosen from the available variables.
}
\description{
Identify the best linear, logistic or mixed regression model using an evolutionary approach.
}
\examples{
data(mtcars)
evolreg(mtcars,"mpg",plot=TRUE)
}
