% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dyn.R
\name{exp_dyn}
\alias{exp_dyn}
\title{Dynamic expansion on a vector (changing the mean without necessarily changing the range of values)}
\usage{
exp_dyn(liste, moyenne, max, round = 0)
}
\arguments{
\item{liste}{A vector}

\item{moyenne}{The new desired average for the vector}

\item{max}{The new max value desired for the vector}

\item{round}{Define the rounding level}
}
\value{
Dynamic expansion on a vector (changing the mean without necessarily changing the range of values)
}
\description{
Dynamic expansion on a vector (changing the mean without necessarily changing the range of values)
}
\examples{
x <- round(rnorm(50,10,4),0) # evaluation of students
hist(x,main="before")
exp_dyn(x,12,20)
hist(x,main="after")
}
