% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.numeric.R
\name{force.numeric}
\alias{force.numeric}
\title{Force numerical variable to be numeric in a dataframe}
\usage{
force.numeric(data)
}
\arguments{
\item{data}{a dataframe.}
}
\value{
The dataframe with numerical variables where it is possible.
}
\description{
Force numerical variable to be numeric in a dataframe
}
\examples{
tablo <- data.frame("A"=as.character(rnorm(20)),"B"=c(rnorm(5),"C3",rnorm(4)),"C"=rep("A",10))
apply(tablo,2,is.numeric)
force.numeric(tablo)-> tablo2 ; apply(tablo2,2,is.numeric)

}
