% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bootstrap.R
\name{identify_ech}
\alias{identify_ech}
\title{Identify the sample size needed to detect a difference between two samples}
\usage{
identify_ech(
  mean1,
  mean2,
  sd1,
  sd2,
  iter = 500,
  conf.level = 0.99,
  conf = 0.95,
  nmin = 2,
  nmax = 1000
)
}
\arguments{
\item{mean1}{Sample Average 1}

\item{mean2}{Sample Average 2}

\item{sd1}{Sample standard deviation 1}

\item{sd2}{Sample standard deviation 2}

\item{iter}{Number of iterations per bootstrap}

\item{conf.level}{Expected confidence on Student's test when comparing 2 samples (0 to 1)}

\item{conf}{Percentage of times I expect to have a p-value that responds to my test if there is an effect when comparing 2 samples}

\item{nmin}{Minimum number of individuals accepted for the sample.}

\item{nmax}{Maximum number of individuals attainable for the sample.}
}
\value{
The required sample size to detect a difference in means by a Student test with the desired significance (\code{conf.level}) and in a minimum number of cases (\code{conf}). If the required sample size exceeds \code{nmax}, a warning message is printed.
}
\description{
Identify the sample size needed to detect a difference between two samples
}
\examples{
identify_ech(mean1=100, mean2=101, sd1=2, sd2=3, nmax=10000)
}
