% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{int.ech}
\alias{int.ech}
\title{Confidence interval of a sample}
\usage{
int.ech(vector, conf.level = 0.95, na.rm = T)
}
\arguments{
\item{vector}{: A vector}

\item{conf.level}{: Confidence level (from 0 to 1)}

\item{na.rm}{: Exclusion of missing values}
}
\value{
Confidence interval on a sample.
}
\description{
Confidence interval of a sample
}
\examples{
x = c(1,4,3,5,3) ; int.ech(x)
}
