% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{int.pop}
\alias{int.pop}
\title{Confidence interval of a population}
\usage{
int.pop(vector, conf.level = 0.95, sigma = c())
}
\arguments{
\item{vector}{: A vector}

\item{conf.level}{: Confidence level (from 0 to 1)}

\item{sigma}{: If known : standard deviation of the population}
}
\value{
Confidence interval on a population.
}
\description{
Confidence interval of a population
}
\examples{
x <- rnorm(1000)
int.pop(x,0.95)
}
