% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{int.prop}
\alias{int.prop}
\title{Confidence interval of a proportion}
\usage{
int.prop(proportion, n, conf.level = 0.95)
}
\arguments{
\item{proportion}{: One or more proportions between 0 and 1.}

\item{n}{: Number of individuals who determined these proportions}

\item{conf.level}{: Confidence level (from 0 to 1)}
}
\value{
Confidence interval on a proportion.
}
\description{
Confidence interval of a proportion
}
\examples{
int.prop(0.50,100,0.95)
#Will give the confidence interval on a proportion of 50\% (0.5) calculated on a sample of 100 items.
}
