% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{int.prop.table}
\alias{int.prop.table}
\title{Confidence intervals on a matrix of a contingent}
\usage{
int.prop.table(x, margin = NULL, conf.level = 0.95)
}
\arguments{
\item{x}{: a matrix (a two-dimensional contingency table: the entries of x must be non-negative integers).}

\item{margin}{: a vector giving the margins to split by. E.g., for a matrix 1 indicates rows, 2 indicates columns, c(1, 2) indicates rows and columns. When x has named dimnames, it can be a character vector selecting dimension names.}

\item{conf.level}{: Confidence level (from 0 to 1)}
}
\value{
Confidence interval on proportions of table obtained with prop.table().
}
\description{
Confidence intervals on a matrix of a contingent
}
\examples{
contingent = with(airquality, table(cut(Temp, quantile(Temp)), Month))
prop.table(contingent) # proportions
int.prop.table(contingent) # confidence intervals
# by row
prop.table(contingent,1) # proportions
int.prop.table(contingent,1) # confidence intervals
# by col
prop.table(contingent,2) # proportions
int.prop.table(contingent,2) # confidence intervals
}
