% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_jb.norm.test.R
\name{jb.norm.test}
\alias{jb.norm.test}
\title{Jarque Bera test for normality}
\usage{
jb.norm.test(x, nrepl = 2000)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{nrepl}{the number of replications in Monte Carlo simulation.}
}
\value{
This function is extracted from the deprecated library normtest

Performs Jarque Bera test for the composite hypothesis of normality, see Jarque and Bera (1987).
}
\description{
Jarque Bera test for normality
}
\examples{
#jb.norm.test(rnorm(100))
}
