% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{kurtosis.norm.test}
\alias{kurtosis.norm.test}
\title{Kurtosis test for normality}
\usage{
kurtosis.norm.test(x, nrepl = 2000)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{nrepl}{the number of replications in Monte Carlo simulation.}
}
\value{
Performs kurtosis test for the composite hypothesis of normality, see, e.g., Shapiro, Wilk and Chen (1968).
}
\description{
Kurtosis test for normality
}
\examples{
# This function is extracted from the deprecated library normtest
#kurtosis.norm.test(rnorm(100))
}
