% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.to.table.R
\name{lms.to.table}
\alias{lms.to.table}
\title{Make a table to summarize coefficients of a list of linear models}
\usage{
lms.to.table(L)
}
\arguments{
\item{L}{A list of linear model objects.}
}
\value{
A dataframe with all the terms of the different linear models and their coefficients.
The minus sign '-' is set for absence of the term in the model.
}
\description{
Make a table to summarize coefficients of a list of linear models
}
\details{
The different linear models must be computed separately, and then
passed to the argument in a list format, using the function list().
}
\examples{
# Simple example
# Assuming lm1, lm2, and lm3 are linear models
# lm_list <- list(lm1, lm2, lm3)
# lms.to.table(L=lm_list)

}
\author{
Julien Bousquet (2021)
}
