% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanbp.R
\name{meanbp}
\alias{meanbp}
\title{Moving average (weighted) per iteration}
\usage{
meanbp(vector, it = 1000, na.rm = T)
}
\arguments{
\item{vector}{: a vector}

\item{it}{: numbre of iterations}

\item{na.rm}{: for ignore missing values}
}
\value{
This function returns an average whose values are weighted according to the distance that separates them from the median initially, and then, at each cycle, with respect to the "average" calculated in the previous cycle.
}
\description{
Moving average (weighted) per iteration
}
