% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.test.R
\name{mm.test}
\alias{mm.test}
\title{Automatic mean/median comparison}
\usage{
mm.test(
  x,
  group,
  pval = 0.05,
  verbose = FALSE,
  return = TRUE,
  paired = FALSE,
  pval_ks = 0.01,
  maxcat = 50,
  plot = TRUE,
  silent = TRUE,
  boot = FALSE,
  iter = 100,
  conf = 0.95,
  code = FALSE,
  debug. = FALSE
)
}
\arguments{
\item{x}{The values: a vector of values, with the samples to be compared.}

\item{group}{Categories: a vector of factors with fewer than maxcat categories.}

\item{pval}{The usual confidence level, 0.05 by default.}

\item{verbose}{\code{FALSE} by default. Increase information.}

\item{return}{Logical. Should the results be returned? Default is \code{TRUE}.}

\item{paired}{\code{FALSE} by default, but can be set to \code{TRUE} if each value of x in categories are paired and in the same order.}

\item{pval_ks}{The p-value for distribution tests like Kolmogorov-Smirnov or Shapiro: 0.01 by default.}

\item{maxcat}{The maximum number of categories: 50 by default.}

\item{plot}{\code{TRUE} to plot the graph with all the results.}

\item{silent}{\code{TRUE} to keep the function silent: no warnings.}

\item{boot}{\code{FALSE} by default. If \code{TRUE}, perform bootstrap sampling with \code{iter} iterations.}

\item{iter}{Number of iterations in the bootstrap, 100 by default.}

\item{conf}{Confidence level of the bootstrap, 0.95 by default.}

\item{code}{If \code{TRUE}, return the code that performs the different tests. \code{FALSE} by default.}

\item{debug.}{\code{FALSE} by default. If \code{TRUE}, print debugging information.}
}
\value{
A list containing the results of the tests.
}
\description{
\code{mm.test()} performs all possible mean comparison tests, even if not mathematically correct and even if hypotheses are wrong. It executes tests to verify hypotheses, and the results are summarized in a graph. The user needs to choose the correct path, which can be done automatically by \code{m.test()}. By default, it finds the path that is more to the left.
}
\examples{
# Example 1
data(iris)
#mm.test(iris$Sepal.Length, iris$Species)
}
\author{
Julien BOUSQUET (2021)
}
