% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Calculation of the mode of a dataset}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
mode
}
\description{
Calculation of the mode of a dataset
}
\examples{
Mode(rnorm(80,20,3))
}
