% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parco.R
\name{parco}
\alias{parco}
\title{Parallel coordinates in a single line of code}
\usage{
parco(data, Y = c(), X = c(), save = F, file = "file.html")
}
\arguments{
\item{data}{A data.frame}

\item{Y}{The Y name of number (for coloration)}

\item{X}{The X list (optional)}

\item{save}{For saving the graph (html format)}

\item{file}{Name of the html page "xxx.html"}
}
\value{
This function allows to explore a large dataset by the parallel coordinates method, using a simplified approach of the plot_ly approach.
}
\description{
Parallel coordinates in a single line of code
}
\examples{
parco(iris,5)
parco(iris,1,2:4)
parco(iris[,1:4],"Sepal.Length")
}
