% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_to_sd.R
\name{range_to_sd}
\alias{range_to_sd}
\title{Estimate standard deviation from a confidence range}
\usage{
range_to_sd(min, max, conf.level = 0.99)
}
\arguments{
\item{min}{Numeric. Lower bound of the observed range.}

\item{max}{Numeric. Upper bound of the observed range.}

\item{conf.level}{Numeric. Confidence level of the range (default = 0.99).}
}
\value{
A numeric value corresponding to the estimated standard deviation.
}
\description{
Computes an estimated standard deviation (\code{sd}) assuming that
the observed range \verb{[min, max]} corresponds to a given confidence level
under a normal distribution.
}
\details{
The function assumes the range represents approximately \verb{±z·σ},
where \code{z} is the quantile of the standard normal distribution
associated with the given confidence level.
}
\examples{
range_to_sd(10, 20, conf.level = 0.95)

}
