% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_rr.R
\name{rr}
\alias{rr}
\title{Gage R&R}
\usage{
rr(Var = c(), optr = c(), app = c(), sigma = 5.15)
}
\arguments{
\item{Var}{measures.}

\item{optr}{identifiers of the three operators.}

\item{app}{identifiers of the 10 pieces.}

\item{sigma}{tolerated variations (multiplication factor of the standard deviation).}
}
\value{
A list containing the following elements:
\describe{
\item{repeatability}{Variability due to the measuring equipment.}
\item{reproducibility}{Variability due to operators.}
\item{RR}{Combined effect of repeatability and reproducibility.}
\item{Vp}{Part - variability due to the measured parts.}
\item{Vt}{Total variability.}
\item{part_Equipement}{Percentage of total variability due to measuring equipment.}
\item{part_Operators}{Percentage of total variability due to operators.}
\item{part_R_R}{Percentage of total variability due to the combination of repeatability and reproducibility.}
\item{part_Pieces}{Percentage of total variability due to the measured parts.}
}
}
\description{
Gage R&R
}
\examples{
set.seed(2)
data <- data.frame("Measures"=rep(rnorm(10,20,0.007),9)*abs(rnorm(90,1,0.025)),
"Operator"=rep(c("A","B","C"),30),
"Equipement"=rep(1:10,9))
boxplot(data$Measures~data$Equipement)
rr(data$Measures,data$Operator,data$Equipement)
}
