% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{skewness.norm.test}
\alias{skewness.norm.test}
\title{Skewness test for normality}
\usage{
skewness.norm.test(x, nrepl = 2000)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{nrepl}{the number of replications in Monte Carlo simulation.}
}
\value{
Performs skewness test for the composite hypothesis of normality, see, e.g., Shapiro, Wilk and Chen (1968).
}
\description{
Skewness test for normality
}
\examples{
# This function is extracted from the deprecated library normtest
#skewness.norm.test(rnorm(100))
}
