% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valreg.R
\name{valreg}
\alias{valreg}
\title{Function to validate a regression model}
\usage{
valreg(reg, analyse = FALSE, nvar = 5)
}
\arguments{
\item{reg}{: A regression model}

\item{analyse}{: To see the detailed balance sheet}

\item{nvar}{: The maximum number of variables allowed.}
}
\value{
This function allows to run all the tests necessary to validate a regression model (check the normal distribution of the residuals, avoid leverage effects, control the variance of the residuals...).
}
\description{
Function to validate a regression model
}
\examples{
#Example 1
data(iris)
reg <- lm(Sepal.Length~.,data=iris[,1:4])
valreg(reg,analyse=TRUE)
#Example 2
data(airquality)
reg <- lm(Wind~Temp,data=airquality)
valreg(reg,analyse=TRUE)
#Example 3
data(mtcars)
reg <- lm(cyl~gear,data=mtcars)
valreg(reg,analyse=TRUE)
#Example 4
data(iris)
reg <- lm(Petal.Width~Petal.Length,data=iris[iris$Species=="virginica",])
valreg(reg,analyse=TRUE)
plot(reg)
}
