% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bootstrap.R
\name{check_ech}
\alias{check_ech}
\title{Determine the differences in mean that I can detect for a known sample size}
\usage{
check_ech(n, mean, sd, conf.level = 0.99, conf = 0.95, iter = 200, cut = 100)
}
\arguments{
\item{n}{: Size of my sample.}

\item{mean}{: Average of my reference sample.}

\item{sd}{: Standard deviation of my sample.}

\item{conf.level}{: Expected confidence on Student's test when I compare 2 samples (0 to 1)}

\item{conf}{: Frequency of times we wish to have the desired p-value defined}

\item{iter}{: Number of iterations per bootstrap}

\item{cut}{: Resolution (number of steps)}
}
\value{
If I have a sample of which I know the size, what effect would I be able to see?


}
\description{
Determine the differences in mean that I can detect for a known sample size
}
\examples{
# What minimum variation could I detect for a sample of 20 individuals (mean 170, sd 20)?
check_ech(n=20,mean = 170,sd = 20)
}
