% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigraph.R
\name{corrigraph}
\alias{corrigraph}
\title{igraph of correlated variables}
\usage{
corrigraph(data, pval = 0.01, exclude = 0.3, ampli = 4)
}
\arguments{
\item{data}{a data.frame}

\item{pval}{the maximum permissible p-value for the display}

\item{exclude}{the minimum threshold of displayed correlations}

\item{ampli}{coefficient of amplification of vertices}
}
\value{
correlation graph network (igraph) of the variables
}
\description{
igraph of correlated variables
}
\examples{
data(swiss)
corrigraph(swiss)
}
