% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{jb.norm.test}
\alias{jb.norm.test}
\title{Jarque-Bera test for normality}
\usage{
jb.norm.test(x, nrepl = 2000)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{nrepl}{the number of replications in Monte Carlo simulation.}
}
\value{
Performs Jarque--Bera test for the composite hypothesis of normality, see Jarque and Bera (1987)..
}
\description{
Jarque-Bera test for normality
}
\examples{
# This function is extracted from the deprecated library normtest
# jb.norm.test(rnorm(100))
# jb.norm.test(abs(runif(100,-2,5)))
}
