% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanbp.R
\name{meanbp}
\alias{meanbp}
\title{Moving average (weighted) per iteration}
\usage{
meanbp(vector, it = 1000, na.rm = T)
}
\arguments{
\item{vector}{: a vector}

\item{it}{: numbre of iterations}

\item{na.rm}{: for ignore missing values}
}
\value{
This function returns an average whose values are weighted according to the distance that separates them from the median initially, and then, at each cycle, with respect to the "average" calculated in the previous cycle.
}
\description{
Moving average (weighted) per iteration
}
\examples{
x <- rnorm(100,20,5) # simulate a sample of a population with an average of 20\%.
meanbp(x,100) # 100 because the calculation will only be repeated a maximum of 100 times.
#if we compare with mean(x) or median(x):
# we see that meanbp gives a result closer to the true average which is 20
}
