% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Calculation of the mode of a dataset}
\usage{
mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
mode
}
\description{
Calculation of the mode of a dataset
}
\examples{
mode(rnorm(80,20,3))
}
