% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.boot.R
\name{pairwise.boot}
\alias{pairwise.boot}
\title{Calculate an pairwise difference between samples by bootstrap without encountering a central-limit problem}
\usage{
pairwise.boot(
  x,
  g,
  iter = 500,
  mu = "meanbp",
  debug = FALSE,
  paired = FALSE,
  verbose = F,
  return = T
)
}
\arguments{
\item{x}{numerical vector}

\item{g}{categorical vector}

\item{iter}{number of iteration}

\item{mu}{comparison criterion, by default 'mean' (arithmetic mean), otherwise 'meanbp' (moving average per iteration: an automatic compromise between mean and median), 'median' or 'sd'.}

\item{debug}{logical, if TRUE enables debug mode for troubleshooting.}

\item{paired}{Logical. If TRUE, performs paired bootstrap tests.}

\item{verbose}{Logical. If TRUE, prints verbose output.}

\item{return}{Logical. If TRUE, returns the results.}
}
\value{
This function returns an array of p-values like the other pairwise functions.

It calculates the number of differences between bootstrap samples that do not include 0 (all higher or all lower).

This number of differences divided by the number of iterations (iter) gives the maximum percentage of times that a convergent difference (all higher or all lower) is found: a confidence.

This value subtracted from 1 gives an equivalent of the p-value whose precision depends on the number of iterations.

Note: using the meanbp criterion is more relevant, it allows a compromise between mean and median by avoiding leverage effects.
}
\description{
Calculate an pairwise difference between samples by bootstrap without encountering a central-limit problem
}
\examples{
# Example 1
data(iris)
pairwise.boot(iris[,2],iris$Species)
# Example 2 by using pairwise(type=="boot")
data(mtcars)
pairwise(mtcars$mpg[mtcars$carb<=4],mtcars$carb[mtcars$carb<=4],type="boot")
}
