% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_rr.R
\name{rr}
\alias{rr}
\title{Gage R&R}
\usage{
rr(Var = c(), optr = c(), app = c(), sigma = 5.15)
}
\arguments{
\item{Var}{: measures.}

\item{optr}{: identifiers of the three operators.}

\item{app}{: identifiers of the 10 pieces.}

\item{sigma}{: tolerated variations (multiplication factor of the standard deviation).}
}
\value{
This function calculates the GageR&R.

It is used to check the reliability of a measuring system.

The following command is used to calculate the following parameters:

repeatability - variability due to the measuring equipment.

reproducibility - variability due to operators.

R&R: Gage - R_R - combined effect of repeatability and reproducibility.

Vp - Part - variability due to the measured parts.

Vt - total variability.

It also directly infers the percentage of the total variability that each of these variabilities represents :

Equipment share: Share_Equipment - Percentage of total variability due to measuring equipment

Operator share: share_Operators - Percentage of total variability due to operators

Share of R&R (\%R&R): share_R_R - Percentage of total variability due to the combination repeatability&reproducibility

Part_Part - Percentage of total variability due to the measured parts
}
\description{
Gage R&R
}
\examples{
set.seed(2)
data <- data.frame("Measures"=rep(rnorm(10,20,0.007),9)*abs(rnorm(90,1,0.025)),
"Operator"=rep(c("A","B","C"),30),
"Equipement"=rep(1:10,9))
boxplot(data$Measures~data$Equipement)
rr(data$Measures,data$Operator,data$Equipement)
}
