% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valreg.R
\name{valreg}
\alias{valreg}
\title{Function to validate a regression model}
\usage{
valreg(
  reg,
  verbose = FALSE,
  nvar = 5,
  boot = FALSE,
  pval = 0.05,
  conf.level = 0.95,
  plot = FALSE,
  raintest_pval = 0.05,
  dwtest_pval = 0.03,
  shapiro_pval = 0.05,
  bptest_pval = 0.05
)
}
\arguments{
\item{reg}{A regression model}

\item{verbose}{To see the detailed balance sheet}

\item{nvar}{The maximum number of variables allowed.}

\item{boot}{For checking the model by bootstrap with bootreg() (FALSE or TRUE)}

\item{pval}{Maximum value accepted for the p-values of the model and its coefficients.}

\item{conf.level}{Confidence interval accepted to validate the regression model by bootstrap.}

\item{plot}{For seeing the graphical analysis of bootreg() (FALSE or TRUE)}

\item{raintest_pval}{Minimal value of p-value accepted for Rainbow test}

\item{dwtest_pval}{Minimal value of p-value accepted for Durbin-Watson test}

\item{shapiro_pval}{Minimal value of p-value accepted for Shapiro-Wilk test}

\item{bptest_pval}{Minimal value of p-value accepted for Breush-Pagan test}
}
\value{
This function allows to run all the tests necessary to validate a regression model (check the normal distribution of the residuals, avoid leverage effects, control the variance of the residuals...).

valreg will therefore validate the regression model, control the p-values and, possibly (boot argument), control the reliability by bootstrap with the bootreg function.
}
\description{
Function to validate a regression model
}
\examples{
#Example 1
data(iris)
reg <- lm(Sepal.Length~.,data=iris[,1:4])
valreg(reg,verbose=TRUE)
#Example 2
data(airquality)
reg <- lm(Wind~Temp,data=airquality)
valreg(reg,verbose=TRUE)
#Example 3
data(mtcars)
reg <- lm(cyl~gear,data=mtcars)
valreg(reg,verbose=TRUE,boot=TRUE)
#Example 4
data(iris)
reg <- lm(Petal.Width~Petal.Length,data=iris[iris$Species=="virginica",])
valreg(reg,verbose=TRUE,boot=TRUE,plot=TRUE)
plot(reg)
#Example 5
data(mtcars);
reg<- lm(cyl~disp+hp,data=mtcars);
corrigraph(mtcars,"cyl");
valreg(reg, verbose=TRUE, plot=TRUE)
}
